Topic_Modeling_Labeling_Single_words
In the earlier version, the corpus was extracted based on the assumption that the folder only contained the speeches’ text files. However, my assumption was wrong and other files got into the corpus along with the speech files. This made us have some words like "file", "corpus", etc. as words occurring in the corpus and as a result turning up in the topic words.
To fix this, the newer version uses a regular expression to extract only the speech files in the folder. So, this version is the correct version since it uses the correct corpus to first Topic Model and then Label.
This version includes the supervised and unsupervised labels produced by the algorithm.